Run prettier stylecheck and change file structure
